requirements.txt
# Lean version of requirements - faster cold start on Streamlit
langchain
langchain-groq
langchain-pinecone
langgraph
langsmith
pinecone
cohere
youtube-transcript-api>=1.2.4
python-dotenv
streamlit
groq